requirements.txt
langchain==0.1.20  
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
pypdf
pandas
rank-bm25
matplotlib
sentence-transformers
chromadb
groq
python-dotenv
